refactor AddressInput using hooks
